// src/app/common-methods/models/cancelRule.ts
export class CancelRule {
    agentId: number;
    agentName: string;
    cancelRuleId: number;
    description: string;
    fromCondition: string;
    fromHours: number;
    amountPercentage: string;
    toCondition: string;
    toHours: number;
    cancelledBy: number;
    active: boolean;
    modifiedBy: number;
    createdBy: number;

}